[tree,html] Fix C++20/GCC13 warnings (enum arithmetic, this capture):
e.g.
```
bitwise operation between different enumeration types ‘TBranch::EStatusBits’ and ‘TBranchElement::EStatusBits’ is deprecated [-Wdeprecated-enum-enum-conversion]
```
```
implicit capture of ‘this’ via ‘[=]’ is deprecated in C++20 [-Wdeprecated]

```

+++ html/src/TDocParser.cxx
@@ -1127,7 +1127,7 @@ Bool_t TDocParser::HandleDirective(TString& line, Ssiz_t& pos, TString& word,
          if (fParseContext.size()>1)
             fParseContext.pop_back();
          if (isInCxxComment && !InContext(kComment)) {
-            fParseContext.push_back(kComment | kCXXComment);
+            fParseContext.push_back((int)kComment | (int)kCXXComment);
             fDocOutput->DecorateEntityBegin(line, pos, kComment);
          }
       }
